remove move doc from redux
