+++ packages/insomnia-app/app/ui/containers/app.tsx
@@ -1578,88 +1578,6 @@ class App extends PureComponent<AppProps, State> {
   }
 }
 
-async function _moveDoc(docToMove, parentId, targetId, targetOffset) {
-  // Nothing to do. We are in the same spot as we started
-  if (docToMove._id === targetId) {
-    return;
-  }
-
-  // Don't allow dragging things into itself or children. This will disconnect
-  // the node from the tree and cause the item to no longer show in the UI.
-  const descendents = await db.withDescendants(docToMove);
-
-  for (const doc of descendents) {
-    if (doc._id === parentId) {
-      return;
-    }
-  }
-
-  function __updateDoc(doc, patch) {
-    // @ts-expect-error -- TSCONVERSION
-    return models.getModel(docToMove.type).update(doc, patch);
-  }
-
-  if (targetId === null) {
-    // We are moving to an empty area. No sorting required
-    await __updateDoc(docToMove, {
-      parentId,
-    });
-    return;
-  }
-
-  // NOTE: using requestToTarget's parentId so we can switch parents!
-  const docs = [
-    ...(await models.request.findByParentId(parentId)),
-    ...(await models.grpcRequest.findByParentId(parentId)),
-    ...(await models.requestGroup.findByParentId(parentId)),
-  ].sort((a, b) => (a.metaSortKey < b.metaSortKey ? -1 : 1));
-
-  // Find the index of doc B so we can re-order and save everything
-  for (let i = 0; i < docs.length; i++) {
-    const doc = docs[i];
-
-    if (doc._id === targetId) {
-      let before, after;
-
-      if (targetOffset < 0) {
-        // We're moving to below
-        before = docs[i];
-        after = docs[i + 1];
-      } else {
-        // We're moving to above
-        before = docs[i - 1];
-        after = docs[i];
-      }
-
-      const beforeKey = before ? before.metaSortKey : docs[0].metaSortKey - 100;
-      const afterKey = after ? after.metaSortKey : docs[docs.length - 1].metaSortKey + 100;
-
-      if (Math.abs(afterKey - beforeKey) < 0.000001) {
-        // If sort keys get too close together, we need to redistribute the list. This is
-        // not performant at all (need to update all siblings in DB), but it is extremely rare
-        // anyway
-        console.log(`[app] Recreating Sort Keys ${beforeKey} ${afterKey}`);
-        await db.bufferChanges(300);
-        docs.map((r, i) =>
-          __updateDoc(r, {
-            metaSortKey: i * 100,
-            parentId,
-          }),
-        );
-      } else {
-        const metaSortKey = afterKey - (afterKey - beforeKey) / 2;
-
-        __updateDoc(docToMove, {
-          metaSortKey,
-          parentId,
-        });
-      }
-
-      break;
-    }
-  }
-}
-
 function mapStateToProps(state: RootState) {
   const { activeActivity, isLoading, loadingRequestIds, isLoggedIn } = state.global;
   // Entities
@@ -1810,7 +1728,6 @@ const mapDispatchToProps = (dispatch: Dispatch<Action<any>>) => {
     handleGoToNextActivity,
     handleExportRequestsToFile,
     handleInitializeEntities,
-    handleMoveDoc: _moveDoc, // TODO this doesn't use dispatch.. it's unclear why it needs to be here.
   };
 };
